Fix: Warnings removed: 1) app_flashled.c - Removed "static uint8_t psz_CRNL[3]= {0x0D, 0x0A, 0};" to remove the variabe computed is not used warning. 2) app_irqs.c - Changed representation of 1000000LL to 1000000L and typecasted it to remove the C99 warning. 3) esos_skel.c - Removed "static uint8_t psz_CRNL[3]= {0x0D, 0x0A, 0};" to remove the variabe computed is not used warning. 4) pic24_serial.c - Added "#include "esos.h"" to remove the implicit declaration of function '__esos_unsafe_PutUint8' warning. - I'm not sure why esos.h isn't included at this point though.

+++ chap14/app_irqs.c
@@ -106,7 +106,7 @@ void  configTimer23(void) {
  *  \return time in microseconds
  */
 uint32_t ticksToUs(uint32_t u32_ticks, uint16_t u16_tmrPre) {
-  return ((uint32_t) ((((uint64_t) u32_ticks) * ((uint64_t) u16_tmrPre) * 1000000LL) / ((uint64_t) FCY)));
+  return ((uint32_t) ((((uint64_t) u32_ticks) * ((uint64_t) u16_tmrPre) * (uint64_t) 1000000L) / ((uint64_t) FCY)));
 }
 
 //Interrupt Service Routine for INT1
